do not trigger the database deleteMedia file, if there is nothing to delete

+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/extensions/Context.kt
@@ -573,7 +573,9 @@ fun Context.getCachedMedia(path: String, getVideosOnly: Boolean = false, getImag
             }
         }
 
-        mediumDao.deleteMedia(*mediaToDelete.toTypedArray())
+        if (mediaToDelete.isNotEmpty()) {
+            mediumDao.deleteMedia(*mediaToDelete.toTypedArray())
+        }
     }.start()
 }
